build: Add the new analysis module
We'll utilise the FileAnalysis struct to probe all of the various
aspects of the installed filesystem so that we might know all required
details ahead of time.

In essence, we'll generate a set of FileAnalysis structs for the entire
installdir, then we'll walk through all the discovered paths and then
automatically assign LayoutEntry's based on the glob matching, instead
of trying to live-match every file encountered.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>