build/manifest: Introduce the new BuildManifest (skeletal)
The BuildManifest type will be used to provide build time verification
that a recipe has been built and tested. Additionally it will be used
to supplement the source index with further information that can only
be known after a recipe has built, such as subpackages, providers of
pkgconfig dependencies, etc.

Relates to #1.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>