build/emitter.d: Create uniqueFiles in 2 steps to ensure only unique entries
Converting `fileSet` to `uniqueFiles` is challenging as it not only filters the entries,
but also transforms the data. What was originally `a.data` becomes `a.hash`. `fileSet`
is not easily sortable by hashes as the hash only exists for some FileTypes. By using an
intermediary `regularFiles`, data is converted to the new struct with complete hashes
making it easy to sort and eliminate duplicates. By using two steps, it ensures there
are no duplicate references in `uniqueFiles` at any stage.

A test of this change on `glibc` with split locale data shows a reduction of 750MB in the
ContentPayload and a 5.5MB reduction in the final package.

Resolves #6

Signed-off-by: Peter O'Connor <sunnyflunk@serpentos.com>