build/profile: Turn preparedFS into an actual preparation of buildRoot
Currently with PGO stages, the sources are extracted over the top of an existing
buildRoot with the last stages build files already present. By making preparedFS
into a logical group, we can recreate buildRoot for every stage of PGO. There's
even some small efficiencies from only running some of the commands when
preparing the buildRoot and not for every stage.

Signed-off-by: Peter O'Connor <sunnyflunk@serpentos.com>