controller: Start on FetchProcessor
This doesn't yet download anything but it does ensure we support a parallel
workflow and awaiting of completion for all upstreams. We'll now need to
actually make this download things, thread safe, and share sources out to
the build itself.

For now we'll set a hard limit of 4 parallel downloads of sources.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>