builder: Use our own collection routine for Collect + Analyse
We request collection target from the mostly skeletal Collector class
and use the Analyser to actually.. analyse.. and bucketise, our remaining
files.

At this point in the changeset we're no longer emitting packages because
the collector is empty, until such point as the remaining code is ported
to Analyser + FileInfo.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>