build/builder: Prevent emission of most Directory LayoutEntrys
This is part of an effort to improve the moss layout to waste as little
as possible while also ensuring efficient unpacks. This change drastically
improves the analysis step time (44ms to 26ms on the nano package on my
system) while also reducing unnecessary layout entries.

We retain the ability to keep empty directories, as well as directories
with special permissions other than 0755.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>