boulder/build: Enforce only /usr/ entries in packages
Currently we accept /usr as a valid entry which can be problematic as we now remove and
later prepend /usr/ onto the path. This allows us to drop a secondary check of whether
the file starts with /usr/ later on (to save a few %). Excludes /usr from printing a
rejection as it's expected and is a rarely hit path so no real cost.

Signed-off-by: Peter O'Connor <sunnyflunk@serpentos.com>