drafter/license: Operate on built-in license files in binary
This means that 1) we don't need to ship the text files and 2) the comparison
is significantly quicker. Lastly, we're getting far more accurate results now
with `amap()` usage though some cases need improving such as openssl detection
which seems hampered by reduced scan length.

This relies heavily on CTFE so the compile times are soaring now.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>