stages/build_package: Implement unconfined builds
This will invoke `mason` through `runuser`, and `fakeroot`,
to ensure we get a dropped pretendy-root environment as we
do in our containerised approach.

However, we share the same view of the filesystem as we would
expect for the "nobody" user, therefore we can perform native
"depless" builds for the purposes of bootstrap.

Limitations: No read permissions outside our own user atm.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>