mason: Reject libtool .la files
These files are generally useless for modern linkers and we can save
disk space by not including them. Not to mention they have a habit of
breaking builds.

In the future we may be forced to provide an option to allow .la
files to exist for any pkgs doing funky things.