boulder new: Add drafter.run() exit status code
This allows for adding extra output in the event of a successful
draft stone.yml creation.

Previously, boulder new would always output the same exit status
regardless of the actual result of the boulder new run.

Still outputs an empty stone.yml in the face of errors accessing
the supplied file, which should probably be considered a bug in
its own right.

Signed-off-by: Rune Morling <ermo@serpentos.com>