boulder/cli: Implement update command to update version in recipes
Updates an existing recipe (stone.yml) with a new version and tarball,
provided by the user. Additionally, it recomputes the hash of the new
tarball and increments the release.

PoC only, to do this properly we need a write function in `moss.source.format`,
that enforces desired formatting. We can already easily parse the recipe.