package.d: Pass "mode=777 to the /dev mount"
Mounting a tmpfs sets the sticky bit by default. When such bit
is set, the kernel prevents from accessing files bind-mounted
inside that tmpfs. In fact, `echo test > /my/container/dev/null`
would fail, when null is a bind-mount of hosts's /dev/null.
"mode=777" does not set the sticky bit and that solves the problem.
See: https://unix.stackexchange.com/questions/619814/why-do-bind-mounts-of-device-nodes-break-with-eacces-in-root-of-a-tmpfs