cli/delete_cache: Use a global counter for getSizeDir
As getSizeDir stats in parallel, there is a race condition when incrementing
the totalSize, and the result will be slightly different every time.

Therefore, make totalSize a global variable and use an atomic operation
to increment it to ensure a consistent result, whilst retaining the speed-up
from doing a parallel stat.

Additionally, turn totalSize into a ulong to avoid an unnecessary cast to
double and return a TLS var.