controller: Tweak search path logic
Introduces configRootPrefix (was configPrefix) and vendorPrefix.

configRootPrefix defaults to "/" and vendorPrefix defaults to a
relativePrefix based on where this boulder instance is run from:

    /usr/local/bin/boulder -> vendorPrefix = /usr/local/bin/../
    /usr/bin/boulder -> vendorPrefix = /usr/bin/../

If a configRootPrefix is specified on the command line, vendorPrefix is
hardcoded to vendorPrefix = /usr to ensure that configurations inside
moss-controlled roots are respected.

configRootPrefix and vendorPrefix are passed along to moss-config, which
then generates config file vendor search paths and admin override search
paths.

Signed-off-by: Rune Morling <ermo@serpentos.com>